requirements.txt
flask
flask-smorest
flask-sqlalchemy